# Simplify erasure code by separating bitrot from erasure code