… cmd/bitrot_test.go
… cmd/erasure-createfile_test.go
… cmd/erasure-healfile_test.go
… cmd/erasure-readfile_test.go
… cmd/erasure.go
… cmd/erasure_test.go
… cmd/posix.go
… cmd/posix_test.go
… cmd/storage-errors.go
… cmd/storage-rpc_test.go
… cmd/xl-v1-common.go
… cmd/xl-v1-healing-common.go
… cmd/xl-v1-healing.go
… cmd/xl-v1-metadata.go
… cmd/xl-v1-multipart.go
… cmd/xl-v1-object.go
… cmd/xl-v1-utils_test.go