<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->
This PR simplifies erasure related code by moving bitrot code out of erasure functions. It started out as a refactor needed to bring in chunked-bitrot feature (i.e instead of one hash for the entire file, have a slice of hashes - one hash per 32MB block), but that feature is on hold as of now. The refactor ended up making the erasure code flow much more straight forward (especially ReadFile)

benchmarks mostly show better performance and consistently less allocs/op
master branch erasure write:
```
goos: linux
goarch: amd64
pkg: github.com/minio/minio/cmd
BenchmarkErasureWriteQuick/_00|00_-4         	     100	  26186484 ns/op	 480.51 MB/s	   11250 B/op	     128 allocs/op
BenchmarkErasureWriteQuick/_00|X0_-4         	     100	  42183467 ns/op	 298.29 MB/s	    8736 B/op	     101 allocs/op
BenchmarkErasureWriteQuick/_X0|00_-4         	     100	  42699211 ns/op	 294.69 MB/s	    8730 B/op	     101 allocs/op
BenchmarkErasureWrite_4_64KB/_00|00_-4       	   10000	    140079 ns/op	 467.85 MB/s	    5895 B/op	      73 allocs/op
BenchmarkErasureWrite_4_64KB/_00|X0_-4       	   10000	    139937 ns/op	 468.32 MB/s	    4901 B/op	      60 allocs/op
BenchmarkErasureWrite_4_64KB/_X0|00_-4       	   10000	    268352 ns/op	 244.22 MB/s	    4902 B/op	      60 allocs/op
BenchmarkErasureWrite_8_20MB/_0000|0000_-4   	      50	 118549636 ns/op	 176.90 MB/s	   22587 B/op	     244 allocs/op
BenchmarkErasureWrite_8_20MB/_0000|X000_-4   	      50	  28331419 ns/op	 740.22 MB/s	   20330 B/op	     218 allocs/op
BenchmarkErasureWrite_8_20MB/_X000|0000_-4   	      50	  38442977 ns/op	 545.52 MB/s	   20345 B/op	     218 allocs/op
BenchmarkErasureWrite_8_20MB/_0000|XXX0_-4   	     100	  50881209 ns/op	 412.17 MB/s	   14981 B/op	     164 allocs/op
BenchmarkErasureWrite_8_20MB/_XXX0|0000_-4   	     100	  61528775 ns/op	 340.84 MB/s	   14976 B/op	     164 allocs/op
BenchmarkErasureWrite_12_30MB/_000000|000000_-4         	      20	 144020088 ns/op	 218.42 MB/s	78695067 B/op	     529 allocs/op
BenchmarkErasureWrite_12_30MB/_000000|X00000_-4         	      30	 122943909 ns/op	 255.87 MB/s	78688434 B/op	     483 allocs/op
BenchmarkErasureWrite_12_30MB/_X00000|000000_-4         	      30	 151870059 ns/op	 207.13 MB/s	78688432 B/op	     483 allocs/op
BenchmarkErasureWrite_12_30MB/_000000|XXXXX0_-4         	      30	 112631694 ns/op	 279.29 MB/s	78675626 B/op	     332 allocs/op
BenchmarkErasureWrite_12_30MB/_XXXXX0|000000_-4         	      30	  50500211 ns/op	 622.91 MB/s	78675625 B/op	     332 allocs/op
BenchmarkErasureWrite_16_40MB/_00000000|00000000_-4     	      30	 222538860 ns/op	 188.48 MB/s	   80554 B/op	     877 allocs/op
BenchmarkErasureWrite_16_40MB/_00000000|X0000000_-4     	      30	 206188438 ns/op	 203.42 MB/s	   76345 B/op	     826 allocs/op
BenchmarkErasureWrite_16_40MB/_X0000000|00000000_-4     	      30	 235677541 ns/op	 177.97 MB/s	   76381 B/op	     827 allocs/op
BenchmarkErasureWrite_16_40MB/_00000000|XXXXXXX0_-4     	      30	  52381580 ns/op	 800.72 MB/s	   51378 B/op	     528 allocs/op
BenchmarkErasureWrite_16_40MB/_XXXXXXX0|00000000_-4     	      30	 105537540 ns/op	 397.42 MB/s	   51386 B/op	     528 allocs/op
PASS
ok  	github.com/minio/minio/cmd	78.872s
```

simplify-erasure branch erasure write:
```
goos: linux
goarch: amd64
pkg: github.com/minio/minio/cmd
BenchmarkErasureWriteQuick/_00|00_-4         	     100	  24848093 ns/op	 506.39 MB/s	   10741 B/op	     120 allocs/op
BenchmarkErasureWriteQuick/_00|X0_-4         	     100	  24434015 ns/op	 514.98 MB/s	    7988 B/op	      92 allocs/op
BenchmarkErasureWriteQuick/_X0|00_-4         	     100	  38507709 ns/op	 326.76 MB/s	    7999 B/op	      92 allocs/op
BenchmarkErasureWrite_4_64KB/_00|00_-4       	   10000	    307006 ns/op	 213.47 MB/s	    5398 B/op	      64 allocs/op
BenchmarkErasureWrite_4_64KB/_00|X0_-4       	   10000	    129687 ns/op	 505.34 MB/s	    4165 B/op	      50 allocs/op
BenchmarkErasureWrite_4_64KB/_X0|00_-4       	   10000	    151819 ns/op	 431.67 MB/s	    4165 B/op	      50 allocs/op
BenchmarkErasureWrite_8_20MB/_0000|0000_-4   	      50	  90487069 ns/op	 231.76 MB/s	   21599 B/op	     228 allocs/op
BenchmarkErasureWrite_8_20MB/_0000|X000_-4   	      50	  52263000 ns/op	 401.27 MB/s	   19079 B/op	     201 allocs/op
BenchmarkErasureWrite_8_20MB/_X000|0000_-4   	      50	  88521639 ns/op	 236.91 MB/s	   19070 B/op	     201 allocs/op
BenchmarkErasureWrite_8_20MB/_0000|XXX0_-4   	      50	  60038124 ns/op	 349.30 MB/s	   14336 B/op	     148 allocs/op
BenchmarkErasureWrite_8_20MB/_XXX0|0000_-4   	     100	  57691554 ns/op	 363.51 MB/s	   13248 B/op	     146 allocs/op
BenchmarkErasureWrite_12_30MB/_000000|000000_-4         	      20	 133012663 ns/op	 236.50 MB/s	78693494 B/op	     506 allocs/op
BenchmarkErasureWrite_12_30MB/_000000|X00000_-4         	      30	 137967677 ns/op	 228.00 MB/s	78686690 B/op	     459 allocs/op
BenchmarkErasureWrite_12_30MB/_X00000|000000_-4         	      30	 115281025 ns/op	 272.87 MB/s	78686738 B/op	     459 allocs/op
BenchmarkErasureWrite_12_30MB/_000000|XXXXX0_-4         	      30	  52844453 ns/op	 595.28 MB/s	78672884 B/op	     304 allocs/op
BenchmarkErasureWrite_12_30MB/_XXXXX0|000000_-4         	      30	  84440031 ns/op	 372.54 MB/s	78672876 B/op	     304 allocs/op
BenchmarkErasureWrite_16_40MB/_00000000|00000000_-4     	      20	 213129602 ns/op	 196.80 MB/s	   82980 B/op	     858 allocs/op
BenchmarkErasureWrite_16_40MB/_00000000|X0000000_-4     	      30	 167005678 ns/op	 251.15 MB/s	   74038 B/op	     796 allocs/op
BenchmarkErasureWrite_16_40MB/_X0000000|00000000_-4     	      20	 227349659 ns/op	 184.49 MB/s	   78505 B/op	     808 allocs/op
BenchmarkErasureWrite_16_40MB/_00000000|XXXXXXX0_-4     	      30	 132088980 ns/op	 317.54 MB/s	   47618 B/op	     491 allocs/op
BenchmarkErasureWrite_16_40MB/_XXXXXXX0|00000000_-4     	      30	 130151638 ns/op	 322.26 MB/s	   47643 B/op	     491 allocs/op
PASS
ok  	github.com/minio/minio/cmd	72.189s
```

master branch erasure-read
```
goos: linux
goarch: amd64
pkg: github.com/minio/minio/cmd
BenchmarkErasureReadQuick/_00|00_-4         	     300	   5198303 ns/op	2420.58 MB/s	16499099 B/op	      57 allocs/op
BenchmarkErasureReadQuick/_00|X0_-4         	     300	   5083063 ns/op	2475.46 MB/s	16512776 B/op	      54 allocs/op
BenchmarkErasureReadQuick/_X0|00_-4         	     200	   8042600 ns/op	1564.53 MB/s	33455470 B/op	      74 allocs/op
BenchmarkErasureReadQuick/_X0|X0_-4         	     200	   8681753 ns/op	1449.35 MB/s	39678510 B/op	      74 allocs/op
BenchmarkErasureRead_4_64KB/_00|00_-4       	   10000	    157246 ns/op	 416.77 MB/s	  754868 B/op	      49 allocs/op
BenchmarkErasureRead_4_64KB/_00|X0_-4       	   10000	    154106 ns/op	 425.26 MB/s	  748092 B/op	      47 allocs/op
BenchmarkErasureRead_4_64KB/_X0|00_-4       	    5000	    205693 ns/op	 318.61 MB/s	  900783 B/op	      56 allocs/op
BenchmarkErasureRead_4_64KB/_X0|X0_-4       	   10000	    214145 ns/op	 306.03 MB/s	  951751 B/op	      56 allocs/op
BenchmarkErasureRead_4_64KB/_00|XX_-4       	   10000	    159268 ns/op	 411.48 MB/s	  756320 B/op	      45 allocs/op
BenchmarkErasureRead_4_64KB/_XX|00_-4       	   10000	    222836 ns/op	 294.10 MB/s	  962344 B/op	      56 allocs/op
BenchmarkErasureRead_8_20MB/_0000|0000_-4   	     200	   7768051 ns/op	2699.71 MB/s	24593491 B/op	     100 allocs/op
BenchmarkErasureRead_8_20MB/_0000|X000_-4   	     200	   7755506 ns/op	2704.08 MB/s	24692850 B/op	      98 allocs/op
BenchmarkErasureRead_8_20MB/_X000|0000_-4   	     100	  10958873 ns/op	1913.66 MB/s	40570543 B/op	     121 allocs/op
BenchmarkErasureRead_8_20MB/_X000|X000_-4   	     100	  11544365 ns/op	1816.60 MB/s	45886715 B/op	     123 allocs/op
BenchmarkErasureRead_8_20MB/_0000|XXXX_-4   	     200	   7721089 ns/op	2716.13 MB/s	24555732 B/op	      92 allocs/op
BenchmarkErasureRead_8_20MB/_XX00|XX00_-4   	     100	  12781197 ns/op	1640.81 MB/s	56225292 B/op	     123 allocs/op
BenchmarkErasureRead_8_20MB/_XXXX|0000_-4   	     100	  14237890 ns/op	1472.94 MB/s	56424589 B/op	     123 allocs/op
BenchmarkErasureRead_12_30MB/_000000|000000_-4         	     100	  11429498 ns/op	2752.29 MB/s	37089924 B/op	     155 allocs/op
BenchmarkErasureRead_12_30MB/_000000|X00000_-4         	     200	  11320750 ns/op	2778.73 MB/s	36988985 B/op	     150 allocs/op
BenchmarkErasureRead_12_30MB/_X00000|000000_-4         	     100	  15015384 ns/op	2095.00 MB/s	53239490 B/op	     184 allocs/op
BenchmarkErasureRead_12_30MB/_X00000|X00000_-4         	     100	  15597873 ns/op	2016.77 MB/s	58490500 B/op	     187 allocs/op
BenchmarkErasureRead_12_30MB/_000000|XXXXXX_-4         	     100	  11651602 ns/op	2699.82 MB/s	37067321 B/op	     143 allocs/op
BenchmarkErasureRead_12_30MB/_XXX000|XXX000_-4         	     100	  18075011 ns/op	1740.37 MB/s	79284403 B/op	     189 allocs/op
BenchmarkErasureRead_12_30MB/_XXXXXX|000000_-4         	      50	  21260468 ns/op	1479.61 MB/s	79569707 B/op	     195 allocs/op
BenchmarkErasureRead_16_40MB/_00000000|00000000_-4     	     100	  14991491 ns/op	2797.79 MB/s	49346079 B/op	     211 allocs/op
BenchmarkErasureRead_16_40MB/_00000000|X0000000_-4     	     100	  14977081 ns/op	2800.48 MB/s	49387808 B/op	     210 allocs/op
BenchmarkErasureRead_16_40MB/_X0000000|00000000_-4     	     100	  18895268 ns/op	2219.76 MB/s	65529347 B/op	     249 allocs/op
BenchmarkErasureRead_16_40MB/_X0000000|X0000000_-4     	     100	  19503575 ns/op	2150.53 MB/s	70709288 B/op	     253 allocs/op
BenchmarkErasureRead_16_40MB/_00000000|XXXXXXXX_-4     	     100	  14994205 ns/op	2797.28 MB/s	49343942 B/op	     195 allocs/op
BenchmarkErasureRead_16_40MB/_XXXX0000|XXXX0000_-4     	      50	  24973150 ns/op	1679.53 MB/s	102262809 B/op	     264 allocs/op
BenchmarkErasureRead_16_40MB/_XXXXXXXX|00000000_-4     	      50	  29751270 ns/op	1409.79 MB/s	102178819 B/op	     263 allocs/op
PASS
ok  	github.com/minio/minio/cmd	65.840s
```

simplify-erasure branch erasure read:
```
goos: linux
goarch: amd64
pkg: github.com/minio/minio/cmd
BenchmarkErasureReadQuick/_00|00_-4         	     300	   4666378 ns/op	2696.50 MB/s	12652586 B/op	      60 allocs/op
BenchmarkErasureReadQuick/_00|X0_-4         	     300	   4661863 ns/op	2699.12 MB/s	12652423 B/op	      57 allocs/op
BenchmarkErasureReadQuick/_X0|00_-4         	     200	   6491840 ns/op	1938.27 MB/s	18944729 B/op	      73 allocs/op
BenchmarkErasureReadQuick/_X0|X0_-4         	     200	   6435338 ns/op	1955.28 MB/s	18944541 B/op	      72 allocs/op
BenchmarkErasureRead_4_64KB/_00|00_-4       	   20000	     74979 ns/op	 874.05 MB/s	  134628 B/op	      54 allocs/op
BenchmarkErasureRead_4_64KB/_00|X0_-4       	   20000	     81826 ns/op	 800.92 MB/s	  134468 B/op	      51 allocs/op
BenchmarkErasureRead_4_64KB/_X0|00_-4       	   10000	    102550 ns/op	 639.06 MB/s	  167584 B/op	      59 allocs/op
BenchmarkErasureRead_4_64KB/_X0|X0_-4       	   10000	    103179 ns/op	 635.16 MB/s	  167440 B/op	      57 allocs/op
BenchmarkErasureRead_4_64KB/_00|XX_-4       	   20000	     74521 ns/op	 879.42 MB/s	  134308 B/op	      48 allocs/op
BenchmarkErasureRead_4_64KB/_XX|00_-4       	   10000	    112764 ns/op	 581.18 MB/s	  200208 B/op	      58 allocs/op
BenchmarkErasureRead_8_20MB/_0000|0000_-4   	     200	   7443535 ns/op	2817.41 MB/s	21110595 B/op	     110 allocs/op
BenchmarkErasureRead_8_20MB/_0000|X000_-4   	     200	   7500163 ns/op	2796.14 MB/s	21110439 B/op	     107 allocs/op
BenchmarkErasureRead_8_20MB/_X000|0000_-4   	     200	   9413000 ns/op	2227.93 MB/s	26354328 B/op	     123 allocs/op
BenchmarkErasureRead_8_20MB/_X000|X000_-4   	     200	   9380155 ns/op	2235.73 MB/s	26354205 B/op	     123 allocs/op
BenchmarkErasureRead_8_20MB/_0000|XXXX_-4   	     200	   7391263 ns/op	2837.34 MB/s	21109954 B/op	      98 allocs/op
BenchmarkErasureRead_8_20MB/_XX00|XX00_-4   	     100	  10476033 ns/op	2001.86 MB/s	31597514 B/op	     122 allocs/op
BenchmarkErasureRead_8_20MB/_XXXX|0000_-4   	     100	  12889754 ns/op	1626.99 MB/s	42083295 B/op	     127 allocs/op
BenchmarkErasureRead_12_30MB/_000000|000000_-4         	     200	  10903475 ns/op	2885.07 MB/s	31715521 B/op	     170 allocs/op
BenchmarkErasureRead_12_30MB/_000000|X00000_-4         	     100	  10937472 ns/op	2876.10 MB/s	31716377 B/op	     170 allocs/op
BenchmarkErasureRead_12_30MB/_X00000|000000_-4         	     100	  13577878 ns/op	2316.80 MB/s	36977670 B/op	     194 allocs/op
BenchmarkErasureRead_12_30MB/_X00000|X00000_-4         	     100	  13449705 ns/op	2338.88 MB/s	36977537 B/op	     194 allocs/op
BenchmarkErasureRead_12_30MB/_000000|XXXXXX_-4         	     100	  11031547 ns/op	2851.57 MB/s	31715563 B/op	     155 allocs/op
BenchmarkErasureRead_12_30MB/_XXX000|XXX000_-4         	     100	  15746882 ns/op	1997.68 MB/s	47495716 B/op	     194 allocs/op
BenchmarkErasureRead_12_30MB/_XXXXXX|000000_-4         	     100	  19749667 ns/op	1592.80 MB/s	63273546 B/op	     204 allocs/op
BenchmarkErasureRead_16_40MB/_00000000|00000000_-4     	     100	  14658585 ns/op	2861.33 MB/s	42223888 B/op	     238 allocs/op
BenchmarkErasureRead_16_40MB/_00000000|X0000000_-4     	     100	  14383188 ns/op	2916.12 MB/s	42223710 B/op	     234 allocs/op
BenchmarkErasureRead_16_40MB/_X0000000|00000000_-4     	     100	  17534808 ns/op	2391.99 MB/s	47469917 B/op	     267 allocs/op
BenchmarkErasureRead_16_40MB/_X0000000|X0000000_-4     	     100	  17492116 ns/op	2397.83 MB/s	47469809 B/op	     268 allocs/op
BenchmarkErasureRead_16_40MB/_00000000|XXXXXXXX_-4     	     100	  14723932 ns/op	2848.63 MB/s	42222593 B/op	     213 allocs/op
BenchmarkErasureRead_16_40MB/_XXXX0000|XXXX0000_-4     	      50	  21861108 ns/op	1918.61 MB/s	63200629 B/op	     277 allocs/op
BenchmarkErasureRead_16_40MB/_XXXXXXXX|00000000_-4     	      50	  28243035 ns/op	1485.08 MB/s	84172121 B/op	     293 allocs/op
PASS
ok  	github.com/minio/minio/cmd	65.308s
```


## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added unit tests to cover my changes.
- [ ] I have added/updated functional tests in [mint](https://github.com/minio/mint). (If yes, add `mint` PR # here: )
- [ ] All new and existing tests passed.