@aead 2018-05-23
@krishnasrinivas Can you also add performance/memory alloc numbers compared to `RELEASE.2018-02-09T22-40-05Z` ? Maybe using [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat)