@krishnasrinivas 2018-05-25
read throughput and allocs/op have significant improvement.

```
krishna@escape:~/go/src/github.com/minio/minio/cmd$ benchstat old.txt new.txt 
name                                        old time/op    new time/op     delta
ErasureWriteQuick/_00|00_-4                   23.2ms ± 0%     23.3ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWriteQuick/_00|X0_-4                   51.5ms ± 0%     37.7ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWriteQuick/_X0|00_-4                   21.7ms ± 0%     32.4ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_4_64KB/_00|00_-4                  162µs ± 0%      136µs ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_4_64KB/_00|X0_-4                  260µs ± 0%      119µs ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_4_64KB/_X0|00_-4                  244µs ± 0%      228µs ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_0000|0000_-4             48.9ms ± 0%     71.1ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_0000|X000_-4             70.5ms ± 0%     77.5ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_X000|0000_-4             97.0ms ± 0%     83.8ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_0000|XXX0_-4             34.0ms ± 0%     75.0ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_XXX0|0000_-4             71.5ms ± 0%     58.2ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_000000|000000_-4         186ms ± 0%      148ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_000000|X00000_-4         180ms ± 0%      125ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_X00000|000000_-4        90.8ms ± 0%    147.5ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_000000|XXXXX0_-4         105ms ± 0%      109ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_XXXXX0|000000_-4        91.8ms ± 0%     92.8ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_00000000|00000000_-4     162ms ± 0%      201ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_00000000|X0000000_-4     214ms ± 0%      220ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_X0000000|00000000_-4     209ms ± 0%      244ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_00000000|XXXXXXX0_-4     140ms ± 0%      140ms ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_XXXXXXX0|00000000_-4     117ms ± 0%      114ms ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_00|00_-4                    6.13ms ± 0%     4.93ms ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_00|X0_-4                    6.09ms ± 0%     4.87ms ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_X0|00_-4                    11.7ms ± 0%      6.7ms ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_X0|X0_-4                    9.00ms ± 0%     6.75ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_00|00_-4                  2.58ms ± 0%     0.13ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_00|X0_-4                  2.55ms ± 0%     0.14ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_X0|00_-4                  2.80ms ± 0%     0.19ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_X0|X0_-4                  2.71ms ± 0%     0.19ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_00|XX_-4                  2.53ms ± 0%     0.14ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_XX|00_-4                  2.57ms ± 0%     0.21ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_0000|0000_-4              8.76ms ± 0%     7.80ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_0000|X000_-4              8.73ms ± 0%     7.85ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_X000|0000_-4              15.1ms ± 0%     10.2ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_X000|X000_-4              13.0ms ± 0%      9.9ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_0000|XXXX_-4              8.86ms ± 0%     8.01ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_XX00|XX00_-4              11.7ms ± 0%     11.1ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_XXXX|0000_-4              12.1ms ± 0%     13.3ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_000000|000000_-4         13.2ms ± 0%     11.6ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_000000|X00000_-4         13.1ms ± 0%     11.6ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_X00000|000000_-4         24.0ms ± 0%     14.3ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_X00000|X00000_-4         21.5ms ± 0%     14.0ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_000000|XXXXXX_-4         13.1ms ± 0%     11.6ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_XXX000|XXX000_-4         16.7ms ± 0%     16.7ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_XXXXXX|000000_-4         18.9ms ± 0%     20.6ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_00000000|00000000_-4     16.1ms ± 0%     15.1ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_00000000|X0000000_-4     16.1ms ± 0%     15.1ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_X0000000|00000000_-4     29.0ms ± 0%     18.2ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_X0000000|X0000000_-4     26.8ms ± 0%     18.1ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_00000000|XXXXXXXX_-4     15.9ms ± 0%     15.3ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_XXXX0000|XXXX0000_-4     22.1ms ± 0%     22.8ms ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_XXXXXXXX|00000000_-4     26.3ms ± 0%     29.2ms ± 0%   ~     (p=1.000 n=1+1)

name                                        old speed      new speed       delta
ErasureWriteQuick/_00|00_-4                  543MB/s ± 0%    539MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWriteQuick/_00|X0_-4                  245MB/s ± 0%    334MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWriteQuick/_X0|00_-4                  580MB/s ± 0%    389MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_4_64KB/_00|00_-4                405MB/s ± 0%    481MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_4_64KB/_00|X0_-4                252MB/s ± 0%    552MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_4_64KB/_X0|00_-4                269MB/s ± 0%    287MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_0000|0000_-4            429MB/s ± 0%    295MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_0000|X000_-4            297MB/s ± 0%    271MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_X000|0000_-4            216MB/s ± 0%    250MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_0000|XXX0_-4            616MB/s ± 0%    280MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_XXX0|0000_-4            293MB/s ± 0%    360MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_000000|000000_-4       169MB/s ± 0%    212MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_000000|X00000_-4       175MB/s ± 0%    251MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_X00000|000000_-4       347MB/s ± 0%    213MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_000000|XXXXX0_-4       300MB/s ± 0%    288MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_XXXXX0|000000_-4       343MB/s ± 0%    339MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_00000000|00000000_-4   259MB/s ± 0%    209MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_00000000|X0000000_-4   196MB/s ± 0%    191MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_X0000000|00000000_-4   201MB/s ± 0%    172MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_00000000|XXXXXXX0_-4   299MB/s ± 0%    300MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_XXXXXXX0|00000000_-4   357MB/s ± 0%    369MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_00|00_-4                  2.05GB/s ± 0%   2.55GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_00|X0_-4                  2.06GB/s ± 0%   2.58GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_X0|00_-4                  1.08GB/s ± 0%   1.86GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_X0|X0_-4                  1.40GB/s ± 0%   1.86GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_00|00_-4                25.4MB/s ± 0%  486.5MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_00|X0_-4                25.7MB/s ± 0%  483.8MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_X0|00_-4                23.4MB/s ± 0%  336.3MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_X0|X0_-4                24.2MB/s ± 0%  350.0MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_00|XX_-4                25.9MB/s ± 0%  482.3MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_XX|00_-4                25.5MB/s ± 0%  310.9MB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_0000|0000_-4            2.40GB/s ± 0%   2.69GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_0000|X000_-4            2.40GB/s ± 0%   2.67GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_X000|0000_-4            1.39GB/s ± 0%   2.06GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_X000|X000_-4            1.61GB/s ± 0%   2.13GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_0000|XXXX_-4            2.37GB/s ± 0%   2.62GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_XX00|XX00_-4            1.80GB/s ± 0%   1.89GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_XXXX|0000_-4            1.74GB/s ± 0%   1.58GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_000000|000000_-4       2.38GB/s ± 0%   2.72GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_000000|X00000_-4       2.40GB/s ± 0%   2.71GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_X00000|000000_-4       1.31GB/s ± 0%   2.20GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_X00000|X00000_-4       1.46GB/s ± 0%   2.25GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_000000|XXXXXX_-4       2.40GB/s ± 0%   2.72GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_XXX000|XXX000_-4       1.88GB/s ± 0%   1.89GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_XXXXXX|000000_-4       1.66GB/s ± 0%   1.53GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_00000000|00000000_-4   2.61GB/s ± 0%   2.77GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_00000000|X0000000_-4   2.60GB/s ± 0%   2.78GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_X0000000|00000000_-4   1.45GB/s ± 0%   2.31GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_X0000000|X0000000_-4   1.57GB/s ± 0%   2.31GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_00000000|XXXXXXXX_-4   2.64GB/s ± 0%   2.75GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_XXXX0000|XXXX0000_-4   1.90GB/s ± 0%   1.84GB/s ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_XXXXXXXX|00000000_-4   1.59GB/s ± 0%   1.44GB/s ± 0%   ~     (p=1.000 n=1+1)

name                                        old alloc/op   new alloc/op    delta
ErasureWriteQuick/_00|00_-4                   11.4kB ± 0%     10.8kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWriteQuick/_00|X0_-4                   9.23kB ± 0%     8.00kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWriteQuick/_X0|00_-4                   9.22kB ± 0%     7.99kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_4_64KB/_00|00_-4                 6.01kB ± 0%     5.40kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_4_64KB/_00|X0_-4                 5.16kB ± 0%     4.17kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_4_64KB/_X0|00_-4                 5.16kB ± 0%     4.17kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_0000|0000_-4             23.0kB ± 0%     21.5kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_0000|X000_-4             21.2kB ± 0%     19.1kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_X000|0000_-4             21.1kB ± 0%     19.1kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_0000|XXX0_-4             16.3kB ± 0%     13.2kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_XXX0|0000_-4             16.3kB ± 0%     13.2kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_000000|000000_-4        78.7MB ± 0%     78.7MB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_000000|X00000_-4        78.7MB ± 0%     78.7MB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_X00000|000000_-4        78.7MB ± 0%     78.7MB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_000000|XXXXX0_-4        78.7MB ± 0%     78.7MB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_XXXXX0|000000_-4        78.7MB ± 0%     78.7MB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_00000000|00000000_-4    82.6kB ± 0%     83.4kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_00000000|X0000000_-4    78.9kB ± 0%     74.1kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_X0000000|00000000_-4    78.9kB ± 0%     74.1kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_00000000|XXXXXXX0_-4    57.4kB ± 0%     47.6kB ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_XXXXXXX0|00000000_-4    57.4kB ± 0%     47.6kB ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_00|00_-4                    22.5MB ± 0%     14.4MB ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_00|X0_-4                    22.5MB ± 0%     14.4MB ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_X0|00_-4                    24.1MB ± 0%     20.8MB ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_X0|X0_-4                    22.9MB ± 0%     20.7MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_00|00_-4                  23.0MB ± 0%      0.6MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_00|X0_-4                  23.0MB ± 0%      0.6MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_X0|00_-4                  24.0MB ± 0%      0.8MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_X0|X0_-4                  23.0MB ± 0%      0.8MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_00|XX_-4                  23.0MB ± 0%      0.6MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_XX|00_-4                  23.0MB ± 0%      0.9MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_0000|0000_-4              24.7MB ± 0%     24.6MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_0000|X000_-4              24.7MB ± 0%     24.6MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_X000|0000_-4              27.5MB ± 0%     29.9MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_X000|X000_-4              26.5MB ± 0%     30.1MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_0000|XXXX_-4              24.7MB ± 0%     24.7MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_XX00|XX00_-4              24.8MB ± 0%     35.3MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_XXXX|0000_-4              24.6MB ± 0%     46.0MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_000000|000000_-4         26.3MB ± 0%     37.0MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_000000|X00000_-4         26.4MB ± 0%     37.0MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_X00000|000000_-4         31.5MB ± 0%     42.5MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_X00000|X00000_-4         30.5MB ± 0%     42.6MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_000000|XXXXXX_-4         26.2MB ± 0%     37.1MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_XXX000|XXX000_-4         26.3MB ± 0%     53.2MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_XXXXXX|000000_-4         26.2MB ± 0%     68.9MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_00000000|00000000_-4     27.9MB ± 0%     49.3MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_00000000|X0000000_-4     27.8MB ± 0%     49.4MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_X0000000|00000000_-4     34.4MB ± 0%     54.7MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_X0000000|X0000000_-4     33.3MB ± 0%     55.0MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_00000000|XXXXXXXX_-4     27.8MB ± 0%     49.4MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_XXXX0000|XXXX0000_-4     27.7MB ± 0%     71.0MB ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_16_40MB/_XXXXXXXX|00000000_-4     27.6MB ± 0%     92.3MB ± 0%   ~     (p=1.000 n=1+1)

name                                        old allocs/op  new allocs/op   delta
ErasureWriteQuick/_00|00_-4                      134 ± 0%        120 ± 0%   ~     (p=1.000 n=1+1)
ErasureWriteQuick/_00|X0_-4                      113 ± 0%         92 ± 0%   ~     (p=1.000 n=1+1)
ErasureWriteQuick/_X0|00_-4                      113 ± 0%         92 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_4_64KB/_00|00_-4                   76.0 ± 0%       64.0 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_4_64KB/_00|X0_-4                   66.0 ± 0%       50.0 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_4_64KB/_X0|00_-4                   66.0 ± 0%       50.0 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_0000|0000_-4                252 ± 0%        228 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_0000|X000_-4                232 ± 0%        201 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_X000|0000_-4                232 ± 0%        201 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_0000|XXX0_-4                190 ± 0%        145 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_8_20MB/_XXX0|0000_-4                190 ± 0%        145 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_000000|000000_-4           544 ± 0%        506 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_000000|X00000_-4           507 ± 0%        459 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_X00000|000000_-4           507 ± 0%        459 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_000000|XXXXX0_-4           386 ± 0%        304 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_12_30MB/_XXXXX0|000000_-4           392 ± 0%        304 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_00000000|00000000_-4       897 ± 0%        859 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_00000000|X0000000_-4       859 ± 0%        796 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_X0000000|00000000_-4       858 ± 0%        796 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_00000000|XXXXXXX0_-4       632 ± 0%        491 ± 0%   ~     (p=1.000 n=1+1)
ErasureWrite_16_40MB/_XXXXXXX0|00000000_-4       632 ± 0%        491 ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_00|00_-4                      93.0 ± 0%       61.0 ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_00|X0_-4                      90.0 ± 0%       58.0 ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_X0|00_-4                       140 ± 0%         75 ± 0%   ~     (p=1.000 n=1+1)
ErasureReadQuick/_X0|X0_-4                       116 ± 0%         74 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_00|00_-4                    66.0 ± 0%       51.0 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_00|X0_-4                    64.0 ± 0%       48.0 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_X0|00_-4                    89.0 ± 0%       56.0 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_X0|X0_-4                    75.0 ± 0%       54.0 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_00|XX_-4                    60.0 ± 0%       45.0 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_4_64KB/_XX|00_-4                    74.0 ± 0%       56.0 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_0000|0000_-4                 175 ± 0%        113 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_0000|X000_-4                 172 ± 0%        110 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_X000|0000_-4                 272 ± 0%        130 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_X000|X000_-4                 251 ± 0%        127 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_0000|XXXX_-4                 163 ± 0%        101 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_XX00|XX00_-4                 207 ± 0%        127 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_8_20MB/_XXXX|0000_-4                 207 ± 0%        134 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_000000|000000_-4            336 ± 0%        177 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_000000|X00000_-4            333 ± 0%        174 ± 0%   ~     (p=1.000 n=1+1)
ErasureRead_12_30MB/_X00000|000000_-4            554 ± 0%        200 ± 0%   ~     (p=1.000 n=1+1)
…